Interface Extensions: refactor API of useCustomProperties so that it works better with memoization + useEffect (#301)
# Summary
https://go/LTT49671

I'm updating the API of `useCustomProperties` to work better with
memoization.

Previously, the API was 
```tsx
useCustomProperties(customProperties: Array<BlockPageElementCustomProperty>): {..}
```

In practice, this was difficult to use. If you didn't memoize
`customProperties` it would be different on every-render, resulting in
us constantly sending messages from the BlockFrame to liveapp to update
the custom properties. If you tried to memoize it, though, it wasn't
easy. For example, suppose you want the following custom properties:
```tsx
       [
         {key: "heatmap_x_axis", label: "X Axis", type: "field", table: base.tables[0]},
         {key: "heatmap_y_axis", label: "Y Axis", type: "field", table: base.tables[0]},
         {
           key: "heatmap_absolute_vs_relative",
           label: "Absolute vs Relative",
           type: "enum",
           possibleValues: [
               { value: "relative", label: "Relative" },
               { value: "absolute", label: "Absolute" },
           ],
           defaultValue: "absolute",
         }
       ]
```

You might try:
```tsx
   const base = useBase();
   const customProperties = useMemo(() => {
       return [ .. ];
     }, [base]);
   const { customPropertyValueByKey } = useCustomProperties(customProperties);
```
However, this doesn't work because `base` will never change. It's a
mutable object instance of the `Base` class. (Yes, the fact that it's
mutable is very un-idiomatic in react.)

If you try
```tsx
   const base = useBase();
   const customProperties = useMemo(() => {
       return [ .. ];
     }, [base.tables]);
   const { customPropertyValueByKey } = useCustomProperties(customProperties);
```
It will be different on every render because `base.tables` actually
computes a new array each time.

If you try
If you try
```tsx
   const base = useBase();
   const customProperties = useMemo(() => {
       return [ .. ];
     }, [base.tables[0]]);
   const { customPropertyValueByKey } = useCustomProperties(customProperties);
```
we have pretty much the same problem as the first attempt.
`base.tables[0]` is a mutable object instance of the `Table` class. It
will change if the table is destroyed or the page element query source
is changed, but won't change if the fields within the table are updated,
for example.

This PR introduces a new API:
```tsx
useCustomProperties(
    getCustomProperties: (base: Base) => Array<BlockPageElementCustomProperty>,
): {..}
```

The return type is the same, but the argument is a function that will
receive `base`. Internally, we set up a subscription to re-run the
`getCustomProperties` function whenever there is a schema change within
`base`.

This can be used as 
```tsx
   const getCustomProperties = useCallback((base) => {
       return [
         {key: "heatmap_x_axis", label: "X Axis", type: "field", table: base.tables[0]},
         {key: "heatmap_y_axis", label: "Y Axis", type: "field", table: base.tables[0]},
         {
           key: "heatmap_absolute_vs_relative",
           label: "Absolute vs Relative",
           type: "enum",
           possibleValues: [
               { value: "relative", label: "Relative" },
               { value: "absolute", label: "Absolute" },
           ],
           defaultValue: "absolute",
         }
       ];
     }, []);
   const { customPropertyValueByKey } = useCustomProperties(getCustomProperties);
```
`getCustomProperties` should still be wrapped in `useCallback`, but in
this example, no dependency array is needed because `base` is passed to
the function.

# Test Plan
Manual