Add optional possibleValues and defaultValue to field property (#303)
# Summary
https://go/LTT49449

Adds optional `possibleValues` and `defaultValue` to fieldId custom
property. This way, a block can filter the fields that are allowed to be
selected. For example, if they have a chart block and want to require
that the y-axis field is a number, they can filter to number fields and
pass `possibleValues`. If `possibleValues` is not provided, all visible
fields in the table will be shown in the dropdown.

Corresponding Hyperbase PR:
https://github.com/Hyperbase/hyperbase/pull/177412

# Test Plan
Manual