Interface Extensions: generate separate docs json files (#308)
# Summary
Generate two docs json files.

One note here is that there is a regression in functionality due to the
way we share code between base blocks and interface blocks. In some
places, the docs now refer to a shared type instead of the specific
class. This happens where the property/method is defined on the shared
class and not overridden by the child class. For example, the
`TableCore` class defines `primaryField`, and this isn't overridden by
the `Table` class in `base/models` or `interface/models` because they
don't need to override it. But as a result, the docs say the type of
`primaryField` is `SdkModeT["FieldT"]` and this does not link anywhere,
whereas before they linked to the `Field` model.

I would like to land this PR even with this shortcoming and aim to
address it in a follow up. As a reminder, there is a manual step to make
these docs reflected live on the site where we copy the json file to
hyperbase. For now, I will only copy the interface blocks json file to
hyperbase and not copy the base blocks json file, so that the existing
docs that are live on airtable.com/developers/extensions/api will not
actually change.

Before:

![image](https://github.com/user-attachments/assets/ab7a66bb-8944-4211-bd41-3f090113736e)

Now: 

![image](https://github.com/user-attachments/assets/e1936f1f-c6e3-4469-be75-7e4d464701a1)


# Test Plan
blocks-docs builds successfully. I copied the resulting json files to
hyperbase and manually verified the documentation.