SDK: React 18/19 support (#304)
# Summary
Related Hyperbase PR: https://github.com/Hyperbase/hyperbase/pull/177371

Allow react 18 and 19 in blocks. The changes in this repo are basically
straightforward.
- Update some versions in package.json files
- Use `react-dom/client` and the `createRoot` API with a fallback to the
old `ReactDOM.render` in initializeBlock
- Noted that the updateBatcher concept in `base/sdk.ts` is not necessary
in react 18+ since updates are batched by default. I left it in place
there for now, but didn't bother adding it to `interface/sdk.ts` (in
fact, I removed some commented code related to it)
- Add some explicit imports to a couple files that use `React` and `JSX`
as types. I think this is needed now because of the [jsx transform
introduced in react
17](https://legacy.reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html),
where you don't have to explicitly import React in order to use JSX.
These files where I had to add the import are files that do *not* use
JSX, but refer to `React` or `React.JSX` as types.

# Test Plan
Manually tested with react 18 and 19, as well as react 16 to make sure
it still works.