Interface Extensions - blocks docs: add post-processing step to docs (#309)
# Summary

(There are some unrelated documentation fixes mixed into this PR as
well, sorry!)
This PR introduces a post processing step to the documentation
generation to accomplish the following 2 things:

### 1. Fixing functionality regression in documentation
This PR addresses the regression in the documentation described in
https://github.com/Hyperbase/blocks-sdk/pull/308. Quoting:
> One note here is that there is a regression in functionality due to
the way we share code between base blocks and interface blocks. In some
places, the docs now refer to a shared type instead of the specific
class. This happens where the property/method is defined on the shared
class and not overridden by the child class. For example, the
`TableCore` class defines `primaryField`, and this isn't overridden by
the `Table` class in `base/models` or `interface/models` because they
don't need to override it. But as a result, the docs say the type of
`primaryField` is `SdkModeT["FieldT"]` and this does not link anywhere,
whereas before they linked to the `Field` model.

This PR addresses this by adding a post-processing step to the
documentation generation, where we replace `SdkModeT["FieldT"]` with the
Field model class found in the json file. Likewise for the Table model.

### 2. Adjusting import paths of shared examples
The post-processing step also includes logic to adjust the import paths
in the examples in doc comments for things that are shared between both
the base and interface documentation. For example, useSynced is
available in both `@airtable/blocks/base/ui` and
`@airtable/blocks/interface/ui`. Previously, the documentation showed an
example of using it as
```ts
import {useSynced} from '@airtable/blocks/base/ui';
```
Now, we can put `@airtable/blocks/[placeholder-path]/ui` in the code
comment and the post-processor will replace that with the correct thing
in each docs json file.

### Other notes
- We could add more post-processing logic in the future.
- We have some existing "custom" logic in the docs generation which is
implemented in some custom typedoc plugins. I chose not to pursue a
typedoc plugin here because 1/ I'd have to learn how the API and it was
easier for me to just write some JSON processing code, and 2/ our
typedoc version is super out of date and there's been tons of breaking
changes, including to the plugin API, so if we want to update typedoc
the plugins will have to be completely rewritten.

# Test Plan
Manually verified. I would like to create some tests for the blocks-docs
package. It doesn't have any right now – probably because it was a
fairly simple wrapper over typedoc, but still worth some tests on the
custom plugins and now post-processing logic. Unfortunately, I'm not
going to fit that in this week and I want to land this to improve the
docs before the interface extensions alpha starts.